requirements.txt
pandas
numpy
automata-lib